requirements.txt
ddg-scraper
disnake
python-dotenv
disnake-jishaku
aiohttp